+++ src/main/java/com/playwithcode/businessbridge/approval/dto/response/DraftListResponse.java
@@ -16,7 +16,7 @@ public class DraftListResponse {
     private final String docForm;                   // 문서 양식
     private final String title;                     // 제목
     private final Integer fileCount;                // 첨부파일
-    private final Long docNo;                       //문서 번호
+    private final int docNo;                        //문서 번호
     private final LocalDateTime draftDateTime;      // 기안일시
 
     public static DraftListResponse from(final Approval approval){
